Refactor App.js
- add useTickerToTable hook
- add Table component
