+++ react-front/src/components/Table.js
@@ -1,0 +1,73 @@
+export default function Table({
+  getTableProps,
+  getTableBodyProps,
+  headerGroups,
+  rows,
+  prepareRow,
+}) {
+  return (
+    <table {...getTableProps()} style={{ border: "solid 1px blue" }}>
+      <thead>
+        {headerGroups.map((headerGroup) => (
+          <tr {...headerGroup.getHeaderGroupProps()}>
+            {headerGroup.headers.map((column) => (
+              <th
+                {...column.getHeaderProps(column.getSortByToggleProps())}
+                style={{
+                  borderBottom: "solid 3px red",
+
+                  background: "aliceblue",
+
+                  color: "black",
+
+                  fontWeight: "bold",
+                }}
+              >
+                {column.render("Header")}
+                <span>
+                  {column.isSorted ? (column.isSortedDesc ? " 🔽" : " 🔼") : ""}
+                </span>
+              </th>
+            ))}
+          </tr>
+        ))}
+      </thead>
+
+      <tbody {...getTableBodyProps()}>
+        {rows.map((row) => {
+          prepareRow(row);
+
+          return (
+            <tr {...row.getRowProps()}>
+              {row.cells.map((cell) => {
+                return (
+                  <td
+                    {...cell.getCellProps()}
+                    style={{
+                      padding: "10px",
+
+                      border: "solid 1px gray",
+
+                      background: "papayawhip",
+                    }}
+                  >
+                    {cell.column.Header === "Ticker" ? (
+                      <a
+                        href={`https://tradingview.com/symbols/${cell.value}`}
+                        target="_blank"
+                      >
+                        {cell.render("Cell")}
+                      </a>
+                    ) : (
+                      cell.render("Cell")
+                    )}
+                  </td>
+                );
+              })}
+            </tr>
+          );
+        })}
+      </tbody>
+    </table>
+  );
+}
